Restructure and fix routing
