+++ frontend/src/AuthenticatedApp/PaperViewPage/subpages/Grading/styles.ts
