+++ frontend/src/AuthenticatedApp/index.tsx
@@ -5,7 +5,7 @@ import { Route, Switch } from "react-router-dom";
 import NotFoundPage from "../App/NotFoundPage";
 import PaperIndex from "./PaperIndexPage";
 import PaperView from "./PaperViewPage";
-import PaperSetup from "./paperSetup/PaperSetup";
+import PaperSetup from "./paperSetup";
 import ScriptView from "./scripts/ScriptView";
 import theme from "./theme";
 
@@ -15,7 +15,7 @@ const AuthenticatedApp: React.FC = () => {
       <CssBaseline />
       <Switch>
         <Route exact path="/" component={PaperIndex} />
-        <Route path="/papers/:paper_id/set_up" component={PaperSetup} />
+        <Route path="/papers/:paper_id/setup" component={PaperSetup} />
         <Route path="/papers/:paper_id" component={PaperView} />
         <Route path="/scripts/:script_id" component={ScriptView} />
         <Route>
